--- version.go
@@ -1,3 +1,0 @@
-package main
-
-var version = "0.0.1"
